requirements.txt
django==3.1.3

selenium==3.141.0
requests==2.25.0

pytz~=2020.4

pyyaml==5.4.1
